fix(trace): extract normalizeToolId to client-safe module, consolidate log-details formatting utils
